generalize buffer reader to any smaller, fixed-length, unsigned type
